users profile

+++ app/Http/Controllers/CategoriesController.php
@@ -91,6 +91,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0)
+        {
+            session()->flash('error', 'Category cannot be deleted because it has some posts..');
+            return redirect()->back();
+
+        }
         $category->delete();
         session()->flash('success', 'Category deleted successfully.');
         return redirect(route('categories.index'));
